Allow an amd64 brlapi.dll to be built using mingw-w64
The mingw-w64 fork of mingw can produce native amd64 Windows binaries,
however it isn't fully compatible with the 32-bit mingw. This commit fixes
some minor compatibility issues, which leads to it being able to build the
client-side brlapi.dll.

Firstly, the `usleep` function that is patched in by BRLTTY for certain
versions of mingw is unnecessary when using mingw-w64, as it is already
exposed by its own libc headers. Therefore, this commit prevents the
compilation of this function if the compiler is detected to be mingw-w64.

Secondly, the `timespec` struct is also already provided when the
mingw-w64 libc is used, so it's similarly excluded.

These two changes are sufficient to get the brlapi.dll itself to compile
(i.e. the `api` make target). The last change makes sure that the
`all-api` target (the tests) can also be built. The LANG_MALAGASY is
undefined when using mingw-w64, so it is conditionally removed. Note that
this language constant is not officially supported [1] on Windows, so
removing it when the target is 64-bit Windows should not be harmful.

[1] List of Language Identifier and Constant Strings
    https://msdn.microsoft.com/en-us/library/windows/desktop/dd318693(v=vs.85).aspx