AtSpi/AtSpi2: default to reading terminal only
Brltty can not really be used alone to read a graphical desktop, Orca is
really needed to do any serious work, and it has much better support for
AtSpi.  Brltty however rocks in terminals, so it does make a lot of
sense to use it along Orca, but only for reading terminals.  It thus
makes sense to make this setup the default.