atspi2: Fix looking for currently active widget
The current code was actually not working at all because the initial
sender is SPI2_DBUS_INTERFACE_REG with path SPI2_DBUS_PATH_ROOT, and its
children are applications with SPI2_DBUS_PATH_ROOT too, so when deciding
to recurse, we need to check for both the path and the sender, not only
the path, which may be the same for two sender and make use believe than
we have a loop right from the beginning between SPI2_DBUS_INTERFACE_REG
and applications.

It is however noticeable that some applications do not actually always
expose their widgets as tree, but sometimes as a clique!  We thus need to
take care not only of direct loops, but also loops to further ancestors,
and just stop recursing completely in that case, to avoid an exponential
number of lookups.