Start API server during initialization
instead of waiting for driver construction. This allows clients, notably
xbrlapi, to know that brltty is present, but the device is not plugged in,
and should thus stay around, ready to start working when the device is
plugged in. When brltty is not present, they will know it from the refused
connexion, and tell the user about it and exit.