brltty-ttb: Add alternated half xcompose files
Some one-handed people are used to having to move their hand between
left and right part. Not having to move the hand is convenient, but they
are used to using the same key for dot1 and dot6, dot2 and dot5, dot3
and dot4, instead of dot1/dot4, dot2/dot4, dot3/dot6.