brlapi: rework suspend functions to make them coherent with resume
api_flush and api_unlink do not need to use a task, and better have the
apiDriverMutex taken in the core thread rather than letting the server keep
it held while waiting for the suspend. That latter point was not posing a
problem yet because suspension does not trigger an api_flush call, but
that could have been possible.