brlapi: Make server have enough buffer for sending/receiving a packet
On Windows, the default named pipe buffer size is actually very small.
We have to tell the OS which minimum size we want. Otherwise we can get
in a situation where both the server and the client are trying to send
data, but none of them is reading each other's output, thus a deadlock.

Making the output buffer large enough allows the server to make sure
that it can write its packets to the client without blocking, and get
back to reading data from the client.