Install systemd units and udev rules correctly
This change makes it so that systemd units and udev rules, as well
as their associated helper scripts, are installed into the correct
directories as needed. We rely on the systemd and udev pkgconfig files
to provide the correct paths and install the helper scripts into
the correct directory for helper binaries.