brlapi: Fix backspace synthesis in xbrlapi
Even if the alternate code without retaindot is the same as the code
with retaindot, we need to call whoGetsKey. Otherwise xbrlapi does not
get backspace keycodes: they produce the same keycode with/without
retaindot, but since xbrlapi sets retaindot to 0, it would not get them
from the first call.