Fixes Debian's bug: https://bugs.debian.org/1007927
Debian adds a custom sysconfig scheme to system python installs,
"posix_local". This is the default scheme, and it redirects local users'
Python module installs to /usr/local even though Python is installed
with a /usr prefix. Both are on Debian's python's sys.path module search
path.

Previously this custom sysconfig scheme was specified in
distutils.sysconfig, but not sysconfig itself. As distutils is being
deprecated, the custom scheme is now specified in sysconfig, since
Debian's Python 3.10 (3.10.2-4).

As far as I can see, pythoncmd is only used in detecting the installed
system python, not determining where the built module will be installed,
so this patch is sufficient. To control where the module is installed,
the approach would be to pass the selected prefix to sysconfig, when
querying the install path.