screen: Set quality to low or none for base and no
When e.g. brltty cannot load a screen driver, but can load the BrlAPI
driver, we have to make sure to know that we have a low screen reading
quality, otherwise the BrlAPI driver would consume braille keyboard
events, without being able to do anything about them.

This notably fixes cursor routing and braille panning in Orca when
xbrlapi is installed but the a2 screen driver is not installed.